refactor tests to test_voting.py
