+++ sklearn/ensemble/tests/test_voting.py
@@ -1,4 +1,4 @@
-"""Testing for the VotingClassifier"""
+"""Testing for the VotingClassifier and AverageRegressor"""
 
 import numpy as np
 from sklearn.utils.testing import assert_almost_equal, assert_array_equal
@@ -18,15 +18,30 @@ from sklearn.datasets import make_multilabel_classification
 from sklearn.svm import SVC
 from sklearn.multiclass import OneVsRestClassifier
 from sklearn.neighbors import KNeighborsClassifier
-from sklearn.base import BaseEstimator, ClassifierMixin
+from sklearn.base import BaseEstimator, ClassifierMixin, RegressorMixin
+from sklearn.ensemble import AverageRegressor
+from sklearn.utils.validation import check_random_state
+from sklearn.model_selection import train_test_split
+from sklearn.linear_model import LinearRegression
+from sklearn.ensemble import RandomForestRegressor
+from sklearn.ensemble import GradientBoostingRegressor
+from sklearn.ensemble import BaggingRegressor
+from sklearn.tree import DecisionTreeRegressor
+from sklearn.neighbors import KNeighborsRegressor
+from sklearn.pipeline import make_pipeline
+from sklearn.feature_selection import SelectKBest
 
+
+# ===========================================================================
+# tests for sklearn.ensemble.VotingClassifier
+# ===========================================================================
 
 # Load the iris dataset and randomly permute it
 iris = datasets.load_iris()
 X, y = iris.data[:, 1:3], iris.target
 
 
-def test_estimator_init():
+def test_estimator_init_voting_classifier():
     eclf = VotingClassifier(estimators=[])
     msg = ('Invalid `estimators` attribute, `estimators` should be'
            ' a list of (string, estimator) tuples')
@@ -65,7 +80,7 @@ def test_predictproba_hardvoting():
     assert_raise_message(AttributeError, msg, eclf.predict_proba, X)
 
 
-def test_notfitted():
+def test_notfitted_voting_classifier():
     eclf = VotingClassifier(estimators=[('lr1', LogisticRegression()),
                                         ('lr2', LogisticRegression())],
                             voting='soft')
@@ -209,7 +224,7 @@ def test_multilabel():
         return
 
 
-def test_gridsearch():
+def test_gridsearch_voting_classifier():
     """Check GridSearch support."""
     clf1 = LogisticRegression(random_state=1)
     clf2 = RandomForestClassifier(random_state=1)
@@ -247,7 +262,7 @@ def test_parallel_fit():
     assert_array_almost_equal(eclf1.predict_proba(X), eclf2.predict_proba(X))
 
 
-def test_sample_weight():
+def test_sample_weight_voting_classifier():
     """Tests sample_weight parameter of VotingClassifier"""
     clf1 = LogisticRegression(random_state=123)
     clf2 = RandomForestClassifier(random_state=123)
@@ -276,7 +291,7 @@ def test_sample_weight():
     assert_raise_message(ValueError, msg, eclf3.fit, X, y, sample_weight)
 
 
-def test_sample_weight_kwargs():
+def test_sample_weight_kwargs_voting_classifier():
     """Check that VotingClassifier passes sample_weight as kwargs"""
     class MockClassifier(BaseEstimator, ClassifierMixin):
         """Mock Classifier to check that sample_weight is received as kwargs"""
@@ -290,7 +305,7 @@ def test_sample_weight_kwargs():
     eclf.fit(X, y, sample_weight=np.ones((len(y),)))
 
 
-def test_set_params():
+def test_set_params_voting_classifier():
     """set_params should be able to set estimators"""
     clf1 = LogisticRegression(random_state=123, C=1.0)
     clf2 = RandomForestClassifier(random_state=123, max_depth=None)
@@ -376,7 +391,7 @@ def test_set_estimator_none():
     assert_array_equal(eclf2.transform(X1), np.array([[0], [1]]))
 
 
-def test_estimator_weights_format():
+def test_estimator_weights_format_voting_classifier():
     # Test estimator weights inputs as list and array
     clf1 = LogisticRegression(random_state=123)
     clf2 = RandomForestClassifier(random_state=123)
@@ -428,3 +443,306 @@ def test_transform():
             eclf3.transform(X).swapaxes(0, 1).reshape((4, 6)),
             eclf2.transform(X)
     )
+
+# ===========================================================================
+# tests for sklearn.ensemble.AverageRegressor
+# ===========================================================================
+
+rng = check_random_state(0)
+# load the boston dataset
+# and randomly permute it
+boston = datasets.load_boston()
+perm = rng.permutation(boston.target.size)
+boston.data = boston.data[perm]
+boston.target = boston.target[perm]
+
+X_train, X_test, y_train, y_test = train_test_split(boston.data[:150],
+                                                    boston.target[:150],
+                                                    random_state=rng)
+
+
+def test_estimator_init_average_regressor():
+    """Test various init configurations"""
+
+    lr = LinearRegression()
+    # test init with empty list of estimators
+    ensemble = AverageRegressor(estimators=[])
+    msg = ('Invalid `estimators` attribute, `estimators` should be'
+           ' a list of (string, estimator) tuples')
+    assert_raise_message(AttributeError, msg, ensemble.fit, X_train, y_train)
+
+    # test init with all estimators set to None
+    ensemble = AverageRegressor(estimators=[('r1', None),
+                                            ('r2', None)])
+    msg = ('All estimators are None. At least one is '
+           'required to be a regressor!')
+    assert_raise_message(ValueError, msg, ensemble.fit, X_train, y_train)
+
+    # test init with shape of estimators different than that of weights
+    ensemble = AverageRegressor(estimators=[('lr', lr)], weights=[1, 2])
+    msg = ('Number of regressors and weights must be equal'
+           '; got 2 weights, 1 estimators')
+    assert_raise_message(ValueError, msg, ensemble.fit, X_train, y_train)
+
+    # valide that estimators names are unique
+    ensemble = AverageRegressor(estimators=[('lr', lr), ('lr', lr)],
+                                weights=[1, 2])
+    msg = "Names provided are not unique: ['lr', 'lr']"
+    assert_raise_message(ValueError, msg, ensemble.fit, X_train, y_train)
+
+    ensemble = AverageRegressor(estimators=[('lr__', lr)])
+    msg = "Estimator names must not contain __: got ['lr__']"
+    assert_raise_message(ValueError, msg, ensemble.fit, X_train, y_train)
+
+    ensemble = AverageRegressor(estimators=[('estimators', lr)])
+    msg = "Estimator names conflict with constructor arguments: ['estimators']"
+    assert_raise_message(ValueError, msg, ensemble.fit, X_train, y_train)
+
+
+def test_notfitted_average_regressor():
+    """Test raise error when predicting with non-fitted estimator"""
+    ensemble = AverageRegressor(estimators=[('lr1', LinearRegression()),
+                                            ('lr2', LinearRegression())])
+    msg = ("This AverageRegressor instance is not fitted yet. Call \'fit\'"
+           " with appropriate arguments before using this method.")
+    assert_raise_message(NotFittedError, msg, ensemble.predict, X_train)
+
+
+def test_single_estimator():
+    """ Check singleton ensembles."""
+
+    clf1 = AverageRegressor(estimators=[
+        ('knn', KNeighborsRegressor())]).fit(X_train, y_train)
+
+    clf2 = KNeighborsRegressor().fit(X_train, y_train)
+
+    assert_array_almost_equal(clf1.predict(X_test), clf2.predict(X_test))
+
+
+def test_average_prediction():
+    """ Check AverageRegressor with multiple models"""
+    reg1 = BaggingRegressor(base_estimator=DecisionTreeRegressor(),
+                            random_state=rng,
+                            n_estimators=100)
+    reg2 = RandomForestRegressor(random_state=rng,
+                                 n_estimators=100)
+    reg3 = GradientBoostingRegressor(random_state=rng,
+                                     n_estimators=100)
+
+    ensemble = AverageRegressor(
+        estimators=[('br', reg1),
+                    ('rf', reg2),
+                    ('gbr', reg3)])
+
+    ensemble.fit(X_train, y_train)
+    scores = cross_val_score(ensemble,
+                             X_train,
+                             y_train,
+                             cv=5,
+                             scoring='r2')
+
+    assert_almost_equal(scores.mean(), 0.85, decimal=1)
+
+
+def test_weights_average_regression_boston():
+    """Check weighted average regression prediction on boston dataset."""
+
+    reg1 = BaggingRegressor(base_estimator=DecisionTreeRegressor(),
+                            random_state=rng,
+                            n_estimators=100)
+    reg2 = RandomForestRegressor(random_state=rng,
+                                 n_estimators=100)
+    reg3 = GradientBoostingRegressor(random_state=rng,
+                                     n_estimators=100)
+
+    ensemble = AverageRegressor(
+        estimators=[('br', reg1),
+                    ('rf', reg2),
+                    ('gbr', reg3)],
+        weights=[1, 2, 10])
+
+    scores = cross_val_score(ensemble,
+                             X_train,
+                             y_train,
+                             cv=5,
+                             scoring='r2')
+    assert_almost_equal(scores.mean(), 0.87, decimal=1)
+
+
+def test_parallel_regression():
+    """ Check parallel regression. """
+
+    estimators = [('dtr', DecisionTreeRegressor())]
+    ensemble = AverageRegressor(estimators=estimators,
+                                n_jobs=3).fit(X_train, y_train)
+
+    ensemble.set_params(n_jobs=1)
+    y1 = ensemble.predict(X_test)
+    ensemble.set_params(n_jobs=2)
+    y2 = ensemble.predict(X_test)
+    assert_array_almost_equal(y1, y2)
+
+    ensemble = AverageRegressor(estimators=estimators,
+                                n_jobs=1).fit(X_train, y_train)
+
+    y3 = ensemble.predict(X_test)
+    assert_array_almost_equal(y1, y3)
+
+
+def test_gridsearch_average_regressor():
+    """Check GridSearch support."""
+
+    reg1 = BaggingRegressor(base_estimator=DecisionTreeRegressor(),
+                            random_state=rng,
+                            n_estimators=10)
+    reg2 = RandomForestRegressor(random_state=rng,
+                                 n_estimators=10)
+    reg3 = GradientBoostingRegressor(random_state=rng,
+                                     n_estimators=10)
+
+    ensemble = AverageRegressor(
+        estimators=[('br', reg1),
+                    ('rf', reg2),
+                    ('gbr', reg3)]
+                    )
+
+    params = {'weights': [[0.5, 0.5, 0.5], [1.0, 0.5, 0.5]]}
+
+    grid = GridSearchCV(estimator=ensemble, param_grid=params, cv=5)
+    grid.fit(X_train, y_train)
+
+
+def test_estimator_weights_format_average_regressor():
+    """ Test estimator weights inputs as list and array """
+
+    reg1 = BaggingRegressor(base_estimator=DecisionTreeRegressor(),
+                            random_state=rng,
+                            n_estimators=10)
+    reg2 = RandomForestRegressor(random_state=rng,
+                                 n_estimators=10)
+    reg3 = GradientBoostingRegressor(random_state=rng,
+                                     n_estimators=10)
+
+    ensemble1 = AverageRegressor(
+        estimators=[('br', reg1),
+                    ('rf', reg2),
+                    ('gbr', reg3)],
+        weights=[1, 2, 10])
+
+    ensemble2 = AverageRegressor(
+        estimators=[('br', reg1),
+                    ('rf', reg2),
+                    ('gbr', reg3)],
+        weights=np.array([1, 2, 10]))
+
+    ensemble1.fit(X_train, y_train)
+    ensemble2.fit(X_train, y_train)
+
+    assert_array_almost_equal(ensemble1.predict(X_test),
+                              ensemble2.predict(X_test))
+
+
+def test_sample_weight_kwargs_average_regressor():
+    """Check that AverageRegressor passes sample_weight as kwargs"""
+    class MockRegressor(BaseEstimator, RegressorMixin):
+        """Mock Regressor to check that sample_weight is received as kwargs"""
+        def fit(self, X, y, *args, **sample_weight):
+            assert_true('sample_weight' in sample_weight)
+
+    clf = MockRegressor()
+    eclf = AverageRegressor(estimators=[('mock', clf)])
+
+    # Should not raise an error.
+    eclf.fit(X_train, y_train, sample_weight=np.ones((len(y_train),)))
+
+
+def test_sample_weight_average_regressor():
+    """Tests sample_weight parameter of AverageRegressor"""
+    reg1 = BaggingRegressor(base_estimator=DecisionTreeRegressor(),
+                            random_state=rng,
+                            n_estimators=10)
+    reg2 = RandomForestRegressor(random_state=rng,
+                                 n_estimators=10)
+    reg3 = GradientBoostingRegressor(random_state=rng,
+                                     n_estimators=10)
+
+    ensemble1 = AverageRegressor(
+        estimators=[('br', reg1),
+                    ('rf', reg2),
+                    ('gbr', reg3)]
+                    ).fit(X_train,
+                          y_train,
+                          sample_weight=np.ones((len(y_train),))
+                          )
+    ensemble2 = AverageRegressor(
+        estimators=[('br', reg1),
+                    ('rf', reg2),
+                    ('gbr', reg3)]
+                    ).fit(X_train,
+                          y_train
+                          )
+
+    assert_array_equal(ensemble1.predict(X_test), ensemble2.predict(X_test))
+
+    sample_weight = np.random.RandomState(123).uniform(size=(len(y_train),))
+
+    ensemble3 = AverageRegressor(
+        estimators=[('gbr', reg3)])
+
+    ensemble3.fit(X_train, y_train, sample_weight)
+    reg3.fit(X_train, y_train, sample_weight)
+    assert_array_equal(ensemble3.predict(X_test), reg3.predict(X_test))
+
+    knnR = KNeighborsRegressor()
+    ensemble4 = AverageRegressor(estimators=[
+        ('rf', reg2), ('knn', knnR)])
+
+    msg = ('Underlying estimator \'knn\' does not support sample weights.')
+    assert_raise_message(
+        ValueError, msg, ensemble4.fit, X_train, y_train, sample_weight)
+
+
+def test_average_regression_with_pipeline():
+
+    pipline = make_pipeline(SelectKBest(k=1), DecisionTreeRegressor())
+    estimator = AverageRegressor([('pp', pipline)])
+
+    estimator.fit(X_train, y_train)
+    assert_true(hasattr(estimator, 'estimators_'))
+
+
+def test_set_params_average_regressor():
+    """set_params should be able to set estimators"""
+    reg1 = LinearRegression()
+    reg2 = RandomForestRegressor(random_state=rng,
+                                 n_estimators=10)
+    reg3 = GradientBoostingRegressor(random_state=rng,
+                                     n_estimators=10)
+
+    ensmble = AverageRegressor(estimators=[
+        ('lr', reg1), ('rf', reg2)], weights=[1, 2])
+
+    assert_true('lr' in ensmble.named_estimators)
+    assert_true(ensmble.named_estimators.lr is ensmble.estimators[0][1])
+    assert_true(ensmble.named_estimators.lr is ensmble.named_estimators['lr'])
+
+    ensmble.fit(X_train, y_train)
+    assert_true('lr' in ensmble.named_estimators_)
+    assert_true(ensmble.named_estimators_.lr is ensmble.estimators_[0])
+    assert_true(ensmble.named_estimators_.lr is
+                ensmble.named_estimators_['lr'])
+
+    ensmble2 = AverageRegressor([
+        ('lr', reg1), ('gbr', reg3)], weights=[1, 2])
+    ensmble2.set_params(gbr=reg2).fit(X_train, y_train)
+    assert_false(hasattr(ensmble2, 'gbr'))
+
+    assert_array_equal(ensmble.predict(X_train), ensmble2.predict(X_train))
+    assert_equal(ensmble2.estimators[0][1].get_params(), reg1.get_params())
+    assert_equal(ensmble2.estimators[1][1].get_params(), reg2.get_params())
+
+    ensmble2.set_params(gbr__max_depth=5)
+    assert_true(ensmble2.estimators[1][1].get_params()['max_depth'] == 5)
+
+    assert_equal(ensmble2.get_params()["gbr__max_depth"],
+                 ensmble2.get_params()["gbr"].get_params()['max_depth'])
